Add multiple dropdowns for easier testing
Having dropdowns at different heights on the page makes it easier to
test all functionality at once. This should result in dropdowns both
going up and down, as well as one that doesn't have enough room for
every entry.

This also lays the groundwork for supporting selecting by class
instead of id -- when we do that, we'll want to test that it applies
correctly to multiple elements.